+++ src/main/java/com/enigmabridge/UserObjectKeyBase.java
@@ -42,7 +42,7 @@ public class UserObjectKeyBase implements UserObjectKey {
     }
 
     @Override
-    public long getUserObjectType() {
+    public int getUserObjectType() {
         return uo.getUserObjectType();
     }
